video-rag-project/requirements.txt
numpy>=1.25.0
pandas>=2.0.0
torch>=2.0.0
faiss-cpu>=1.7.4
transformers>=4.30.0
tqdm>=4.65.0
